requirements.txt

streamlit
pandas
google-cloud-firestore
google-auth
google-auth-oauthlib
google-auth-httplib2
protobuf==4.25.3
